Refactor common presenter methods
refactor similar sections rendering
